+++ packages/zero-client/src/client/custom.test.ts
@@ -1,17 +1,17 @@
 import {beforeEach, describe, expect, expectTypeOf, test} from 'vitest';
 import {schema} from '../../../zql/src/query/test/test-schemas.ts';
-import {
-  TransactionImpl,
-  type CustomMutatorDefs,
-  type MakeCustomMutatorInterfaces,
-  type Transaction,
-} from './custom.ts';
+import {TransactionImpl} from './custom.ts';
 import {zeroForTest} from './test-utils.ts';
 import {nanoid} from '../util/nanoid.ts';
 import {createDb} from './test/create-db.ts';
 import {IVMSourceRepo} from './ivm-source-repo.ts';
 import type {WriteTransaction} from './replicache-types.ts';
 import {must} from '../../../shared/src/must.ts';
+import type {
+  CustomMutatorDefs,
+  MakeCustomMutatorInterfaces,
+  Transaction,
+} from '../../../zql/src/mutate/custom.ts';
 type Schema = typeof schema;
 
 test('argument types are preserved on the generated mutator interface', () => {
